test/fixtures/wpt/resource-timing/entry-attributes.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing: PerformanceResourceTiming attributes</title>
<link rel="author" title="Google" href="http://www.google.com/" />
<link rel="help"
  href="https://www.w3.org/TR/resource-timing-2/#timing-allow-origin"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/resource-loaders.js"></script>
<script src="resources/entry-invariants.js"></script>
<script>
attribute_test(
  load.image, "resources/fake_responses.py#hash=1",
  entry => {
    assert_true(entry.name.includes('#hash=1'),
      "There should be a hash in the resource name");
    invariants.assert_tao_pass_no_redirect_http(entry);
  },
  "Image resources should generate conformant entries");

attribute_test(
  load.font, "/fonts/Ahem.ttf",
  invariants.assert_tao_pass_no_redirect_http,
  "Font resources should generate conformant entries");

attribute_test(
  load.image, "/common/redirect.py?location=resources/fake_responses.py",
  invariants.assert_same_origin_redirected_resource,
  "Same-origin redirects should populate redirectStart/redirectEnd");
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that PerformanceResourceTiming entries' attributes are
populated with the correct values.</p>
</body>
</html>
